Renaming JWTSuport to jwt
